requirements.txt
botbuilder-core>=4.7.0
aiohttp
Flask
tabulate
openai
azure-core